dashboard/requirements.txt
nicegui>=2.19.0
requests>=2.32.0
kubernetes>=28.1.0
python-dotenv>=1.0.0
paramiko>=3.4.0
python-dateutil>=2.8.2
